Restructred getCompareListdata function into multiple smaller functions
